move menu classes into right packages
