+++ SmartWarehouseWeb/src/main/java/pl/infoshare/controller/MenuObjectController.java
@@ -1,17 +1,18 @@
-package pl.infoshare.menu;
+package pl.infoshare.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import pl.infoshare.service.MenuObjectService;
 
 @Controller
 @RequestMapping
-public class MenuController {
+public class MenuObjectController {
 
-    private final MenuService menuService;
+    private final MenuObjectService menuService;
 
-    public MenuController(MenuService menuService) {
+    public MenuObjectController(MenuObjectService menuService) {
         this.menuService = menuService;
     }
